Read arguments from argv in bin and pass them to create
