+++ src/bin.mjs
@@ -1,10 +1,14 @@
 #!/usr/bin/env node
 
 import { create } from "./main.mjs";
-import { debug } from "./util.mjs";
+import { debug, deriveProjectNameAndPath } from "./util.mjs";
+import { argv } from "node:process";
 
 try {
-  await create();
+  const { projectName, projectPath } = deriveProjectNameAndPath(argv[2]);
+  debug("projectName", projectName);
+  debug("projectPath", projectPath);
+  await create({ projectName, projectPath });
 } catch (err) {
   console.error("create-ts-node:", err.message);
   debug(err.stack);
